+++ format/format_test.go
@@ -1,4 +1,4 @@
-package promfmt
+package format
 
 import (
 	"testing"
